+++ ztb/analysis/__init__.py
@@ -5,5 +5,10 @@
 """
 
 from .core.analyzer import UnifiedAnalyzer
+from .market_regime_types import MarketRegime, RegimeDetectionResult
 
-__all__ = ["UnifiedAnalyzer"]
+__all__ = [
+    "UnifiedAnalyzer",
+    "MarketRegime",
+    "RegimeDetectionResult"
+]